requirements.txt
streamlit==1.37.1
pyyaml==6.0.2
rapidfuzz==3.9.6